requirements.txt
boto3==1.34.148
pytest==8.3.2
python-dotenv==1.0.1
